requirements.txt.txt
streamlit
ultralytics
onnxruntime   # needed if you want to run best.onnx
#torch         # needed if you want to run best.torchscript
pillow
